refactor: consolidate form schema and action handling in user authentication
