+++ app/routes/_auth+/sign-in/components/user-auth-form.tsx
@@ -9,8 +9,7 @@ import { Button } from '~/components/ui/button'
 import { Input } from '~/components/ui/input'
 import { Label } from '~/components/ui/label'
 import { cn } from '~/lib/utils'
-import type { action } from '../route'
-import { formSchema } from '../schema'
+import { type action, formSchema } from '../route'
 
 type UserAuthFormProps = HTMLAttributes<HTMLFormElement>
